+++ SplunkCrashReports/Sources/SplunkCrashReports/CrashReports.swift
@@ -42,6 +42,8 @@ public class CrashReports {
 
     private var crashReporter: PLCrashReporter?
 
+    var crashSpanName: String = "SplunkCrashReport"
+
     /// Storage of periodically sampled device data.
     private var deviceDataDictionary: [String: String] = [:]
     private var dataUpdateTimer: Timer?
@@ -151,8 +153,6 @@ public class CrashReports {
         }
     }
 
-    // MARK: - Private methods
-
     /// Starts up crash reporter if enable is true and no debugger attached.
     public func initializeCrashReporter() -> Bool {
 
@@ -191,6 +191,20 @@ public class CrashReports {
         startPollingForDeviceStats()
 
         return true
+    }
+
+    // MARK: - Private methods
+
+
+    /// Updates the span name with report data.
+    ///
+    /// Only updates if the provided name is not empty.
+    func updateSpanName(_ name: String) {
+        guard !name.isEmpty else {
+            return
+        }
+
+        crashSpanName = name
     }
 
     /// Returns true if debugger is attached.
@@ -282,7 +296,7 @@ public class CrashReports {
                 instrumentationVersion: sharedState?.agentVersion
             )
 
-        let crashSpan = tracer.spanBuilder(spanName: "SplunkCrashReport")
+        let crashSpan = tracer.spanBuilder(spanName: crashSpanName)
             .setStartTime(time: timestamp)
             .startSpan()